backend/HouseholdProject/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==2.9.2
cycler==0.10.0
Django==3.1.3
django-auth0==0.0.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-jwt==1.13.4
idna==2.10
isort==5.4.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.2
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.19.1
Pillow==8.2.0
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
rawpy==0.16.0
requests==2.25.1
scipy==1.5.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.1
urllib3==1.26.3
wrapt==1.12.1
zombiedice==0.1.5
